Update main, add setup instructions and requirements
