+++ .gitignore
@@ -9,3 +9,13 @@ src/__pycache__/io.cpython-36.pyc
 src/__pycache__/loading.cpython-36.pyc
 src/__pycache__/processing.cpython-36.pyc
 src/__pycache__/loading.cpython-36.pyc
+output.json
+src/__pycache__/__main__.cpython-36.pyc
+src/__pycache__/loading.cpython-36.pyc
+unbabel_cli.egg-info/dependency_links.txt
+unbabel_cli.egg-info/entry_points.txt
+unbabel_cli.egg-info/PKG-INFO
+unbabel_cli.egg-info/requires.txt
+unbabel_cli.egg-info/SOURCES.txt
+unbabel_cli.egg-info/top_level.txt
+src/__pycache__/loading.cpython-36.pyc
