+++ requirements.txt
@@ -1,0 +1,3 @@
+pytest==4.5.0
+numpy==1.17.2
+pandas==0.23.4
